Flytter filer til mer korrekte pakker
