--- src/main/kotlin/no/nav/familie/klage/brev/FormkravVilkår.kt
@@ -1,3 +1,0 @@
-package no.nav.familie.klage.brev
-
-interface FormkravVilkår
