+++ src/main/kotlin/no/nav/familie/klage/brev/baks/FritekstbrevHtmlUtleder.kt
@@ -1,8 +1,8 @@
 package no.nav.familie.klage.brev.baks
 
 import no.nav.familie.klage.behandling.domain.Behandling
-import no.nav.familie.klage.brev.BrevClient
-import no.nav.familie.klage.brev.BrevsignaturService
+import no.nav.familie.klage.brev.felles.BrevClient
+import no.nav.familie.klage.brev.felles.BrevsignaturService
 import no.nav.familie.klage.fagsak.FagsakService
 import no.nav.familie.klage.personopplysninger.PersonopplysningerService
 import org.springframework.stereotype.Component
