+++ src/main/kotlin/no/nav/familie/klage/brev/ef/BrevService.kt
@@ -6,20 +6,15 @@ import no.nav.familie.klage.behandling.domain.PåklagetVedtakDetaljer
 import no.nav.familie.klage.behandling.domain.PåklagetVedtakstype
 import no.nav.familie.klage.behandling.domain.StegType
 import no.nav.familie.klage.behandling.domain.erLåstForVidereBehandling
-import no.nav.familie.klage.brev.BrevClient
-import no.nav.familie.klage.brev.BrevInnholdUtleder
-import no.nav.familie.klage.brev.BrevsignaturService
-import no.nav.familie.klage.brev.FamilieDokumentClient
-import no.nav.familie.klage.brev.FritekstBrevRequestDto
-import no.nav.familie.klage.brev.ef.BrevmottakerUtil.validerMinimumEnMottaker
-import no.nav.familie.klage.brev.ef.BrevmottakerUtil.validerUnikeBrevmottakere
-import no.nav.familie.klage.brev.ef.domain.Brev
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerPerson
-import no.nav.familie.klage.brev.ef.domain.Brevmottakere
-import no.nav.familie.klage.brev.ef.domain.BrevmottakereJournalposter
-import no.nav.familie.klage.brev.ef.domain.MottakerRolle
-import no.nav.familie.klage.brev.ef.dto.BrevmottakereDto
-import no.nav.familie.klage.brev.ef.dto.tilDomene
+import no.nav.familie.klage.brev.ef.brevmottaker.BrevmottakerUtil.validerMinimumEnMottaker
+import no.nav.familie.klage.brev.ef.brevmottaker.BrevmottakerUtil.validerUnikeBrevmottakere
+import no.nav.familie.klage.brev.ef.brevmottaker.BrevmottakereDto
+import no.nav.familie.klage.brev.ef.brevmottaker.tilDomene
+import no.nav.familie.klage.brev.felles.BrevClient
+import no.nav.familie.klage.brev.felles.BrevInnholdUtleder
+import no.nav.familie.klage.brev.felles.BrevsignaturService
+import no.nav.familie.klage.brev.felles.FamilieDokumentClient
+import no.nav.familie.klage.brev.felles.FritekstBrevRequestDto
 import no.nav.familie.klage.fagsak.FagsakService
 import no.nav.familie.klage.fagsak.domain.Fagsak
 import no.nav.familie.klage.felles.domain.Fil
