+++ src/main/kotlin/no/nav/familie/klage/brev/ef/brevmottaker/BrevmottakereDto.kt
@@ -1,8 +1,8 @@
-package no.nav.familie.klage.brev.ef.dto
+package no.nav.familie.klage.brev.ef.brevmottaker
 
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerOrganisasjon
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerPerson
-import no.nav.familie.klage.brev.ef.domain.Brevmottakere
+import no.nav.familie.klage.brev.ef.BrevmottakerOrganisasjon
+import no.nav.familie.klage.brev.ef.BrevmottakerPerson
+import no.nav.familie.klage.brev.ef.Brevmottakere
 
 data class BrevmottakereDto(
     val personer: List<BrevmottakerPerson>,
