+++ src/main/kotlin/no/nav/familie/klage/brev/felles/BrevInnholdUtleder.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.klage.brev
+package no.nav.familie.klage.brev.felles
 
 import no.nav.familie.klage.behandling.domain.PåklagetVedtakDetaljer
 import no.nav.familie.klage.felles.util.StønadstypeVisningsnavn.visningsnavn
@@ -27,30 +27,30 @@ class BrevInnholdUtleder(
             navn = navn,
             personIdent = ident,
             avsnitt =
-            listOf(
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold =
-                    "Vi har ${klageMottatt.norskFormat()} fått klagen din på vedtaket om " +
-                        "${visningsnavn(stønadstype, påklagetVedtakDetaljer)} som ble gjort " +
-                        "${påklagetVedtakDetaljer.vedtakstidspunkt.norskFormat()}, " +
-                        "og kommet frem til at vi ikke endrer vedtaket. Nav Klageinstans skal derfor vurdere saken din på nytt.",
+                listOf(
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold =
+                            "Vi har ${klageMottatt.norskFormat()} fått klagen din på vedtaket om " +
+                                "${visningsnavn(stønadstype, påklagetVedtakDetaljer)} som ble gjort " +
+                                "${påklagetVedtakDetaljer.vedtakstidspunkt.norskFormat()}, " +
+                                "og kommet frem til at vi ikke endrer vedtaket. Nav Klageinstans skal derfor vurdere saken din på nytt.",
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = "Saksbehandlingstidene finner du på nav.no/saksbehandlingstider.",
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "Dette er vurderingen vi har sendt til Nav Klageinstans",
+                        innhold = instillingKlageinstans,
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "Har du nye opplysninger?",
+                        innhold =
+                            "Har du nye opplysninger eller ønsker å uttale deg, kan du sende oss dette via \n${stønadstype.klageUrl()}.",
+                    ),
+                    harDuSpørsmålAvsnitt(stønadstype),
                 ),
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = "Saksbehandlingstidene finner du på nav.no/saksbehandlingstider.",
-                ),
-                AvsnittDto(
-                    deloverskrift = "Dette er vurderingen vi har sendt til Nav Klageinstans",
-                    innhold = instillingKlageinstans,
-                ),
-                AvsnittDto(
-                    deloverskrift = "Har du nye opplysninger?",
-                    innhold =
-                    "Har du nye opplysninger eller ønsker å uttale deg, kan du sende oss dette via \n${stønadstype.klageUrl()}.",
-                ),
-                harDuSpørsmålAvsnitt(stønadstype),
-            ),
         )
 
     fun lagFormkravAvvistBrev(
@@ -69,27 +69,27 @@ class BrevInnholdUtleder(
             personIdent = ident,
             navn = navn,
             avsnitt =
-            listOf(
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = avvistBrevInnhold.årsakTilAvvisning,
+                listOf(
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = avvistBrevInnhold.årsakTilAvvisning,
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = avvistBrevInnhold.brevtekstFraSaksbehandler,
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = avvistBrevInnhold.lovtekst,
+                    ),
+                    duHarRettTilÅKlageAvsnitt(stønadstype),
+                    AvsnittDto(
+                        deloverskrift = "Du har rett til innsyn",
+                        innhold =
+                            "På nav.no/dittnav kan du se dokumentene i saken din.",
+                    ),
+                    harDuSpørsmålAvsnitt(stønadstype),
                 ),
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = avvistBrevInnhold.brevtekstFraSaksbehandler,
-                ),
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = avvistBrevInnhold.lovtekst,
-                ),
-                duHarRettTilÅKlageAvsnitt(stønadstype),
-                AvsnittDto(
-                    deloverskrift = "Du har rett til innsyn",
-                    innhold =
-                    "På nav.no/dittnav kan du se dokumentene i saken din.",
-                ),
-                harDuSpørsmålAvsnitt(stønadstype),
-            ),
         )
     }
 
@@ -107,27 +107,27 @@ class BrevInnholdUtleder(
             personIdent = ident,
             navn = navn,
             avsnitt =
-            listOf(
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = "Vi har avvist klagen din fordi du ikke har klaget på et vedtak.",
+                listOf(
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = "Vi har avvist klagen din fordi du ikke har klaget på et vedtak.",
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = brevtekstFraSaksbehandler,
+                    ),
+                    AvsnittDto(
+                        deloverskrift = "",
+                        innhold = "Vedtaket er gjort etter forvaltningsloven §§ 28 og 33.",
+                    ),
+                    duHarRettTilÅKlageAvsnitt(stønadstype),
+                    AvsnittDto(
+                        deloverskrift = "Du har rett til innsyn",
+                        innhold =
+                            "På nav.no/dittnav kan du se dokumentene i saken din.",
+                    ),
+                    harDuSpørsmålAvsnitt(stønadstype),
                 ),
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = brevtekstFraSaksbehandler,
-                ),
-                AvsnittDto(
-                    deloverskrift = "",
-                    innhold = "Vedtaket er gjort etter forvaltningsloven §§ 28 og 33.",
-                ),
-                duHarRettTilÅKlageAvsnitt(stønadstype),
-                AvsnittDto(
-                    deloverskrift = "Du har rett til innsyn",
-                    innhold =
-                    "På nav.no/dittnav kan du se dokumentene i saken din.",
-                ),
-                harDuSpørsmålAvsnitt(stønadstype),
-            ),
         )
     }
 
@@ -135,17 +135,17 @@ class BrevInnholdUtleder(
         AvsnittDto(
             deloverskrift = "Du har rett til å klage",
             innhold =
-            "Hvis du vil klage, må du gjøre dette innen 6 uker fra den datoen du fikk dette brevet. " +
-                "Du finner skjema og informasjon på ${stønadstype.klageUrl()}.",
+                "Hvis du vil klage, må du gjøre dette innen 6 uker fra den datoen du fikk dette brevet. " +
+                    "Du finner skjema og informasjon på ${stønadstype.klageUrl()}.",
         )
 
     private fun harDuSpørsmålAvsnitt(stønadstype: Stønadstype) =
         AvsnittDto(
             deloverskrift = "Har du spørsmål?",
             innhold =
-            "Du finner mer informasjon på ${stønadstype.lesMerUrl()}.\n\n" +
-                "På nav.no/kontakt kan du chatte eller skrive til oss.\n\n" +
-                "Hvis du ikke finner svar på nav.no kan du ringe oss på telefon 55 55 33 33, hverdager 09.00-15.00.",
+                "Du finner mer informasjon på ${stønadstype.lesMerUrl()}.\n\n" +
+                    "På nav.no/kontakt kan du chatte eller skrive til oss.\n\n" +
+                    "Hvis du ikke finner svar på nav.no kan du ringe oss på telefon 55 55 33 33, hverdager 09.00-15.00.",
         )
 
     private fun visningsnavn(
