+++ src/main/kotlin/no/nav/familie/klage/brev/felles/FormkravVilkår.kt
@@ -1,0 +1,3 @@
+package no.nav.familie.klage.brev.felles
+
+interface FormkravVilkår
