+++ src/main/kotlin/no/nav/familie/klage/brev/felles/SignaturDto.kt
@@ -1,3 +1,3 @@
-package no.nav.familie.klage.brev
+package no.nav.familie.klage.brev.felles
 
 data class SignaturDto(val navn: String, val enhet: String)
