+++ src/main/kotlin/no/nav/familie/klage/distribusjon/ef/JournalførBrevTask.kt
@@ -1,12 +1,12 @@
 package no.nav.familie.klage.distribusjon.ef
 
 import no.nav.familie.klage.behandling.BehandlingService
+import no.nav.familie.klage.brev.ef.Brev
 import no.nav.familie.klage.brev.ef.BrevService
-import no.nav.familie.klage.brev.ef.BrevmottakerUtil.validerMinimumEnMottaker
-import no.nav.familie.klage.brev.ef.domain.Brev
-import no.nav.familie.klage.brev.ef.domain.Brevmottakere
-import no.nav.familie.klage.brev.ef.domain.BrevmottakereJournalpost
-import no.nav.familie.klage.brev.ef.domain.BrevmottakereJournalposter
+import no.nav.familie.klage.brev.ef.Brevmottakere
+import no.nav.familie.klage.brev.ef.BrevmottakereJournalpost
+import no.nav.familie.klage.brev.ef.BrevmottakereJournalposter
+import no.nav.familie.klage.brev.ef.brevmottaker.BrevmottakerUtil.validerMinimumEnMottaker
 import no.nav.familie.klage.distribusjon.DistribusjonService
 import no.nav.familie.klage.distribusjon.SendTilKabalTask
 import no.nav.familie.klage.distribusjon.ef.JournalføringUtil.mapAvsenderMottaker
