+++ src/main/kotlin/no/nav/familie/klage/infrastruktur/config/DatabaseConfiguration.kt
@@ -3,8 +3,8 @@ package no.nav.familie.klage.infrastruktur.config
 import com.fasterxml.jackson.module.kotlin.readValue
 import no.nav.familie.klage.behandling.domain.FagsystemRevurdering
 import no.nav.familie.klage.behandling.domain.PåklagetVedtakDetaljer
-import no.nav.familie.klage.brev.ef.domain.Brevmottakere
-import no.nav.familie.klage.brev.ef.domain.BrevmottakereJournalposter
+import no.nav.familie.klage.brev.ef.Brevmottakere
+import no.nav.familie.klage.brev.ef.BrevmottakereJournalposter
 import no.nav.familie.klage.felles.domain.Endret
 import no.nav.familie.klage.felles.domain.Fil
 import no.nav.familie.kontrakter.felles.objectMapper
