+++ src/main/kotlin/no/nav/familie/klage/kabal/KabalService.kt
@@ -2,11 +2,11 @@ package no.nav.familie.klage.kabal
 
 import no.nav.familie.klage.behandling.domain.Behandling
 import no.nav.familie.klage.behandling.domain.PåklagetVedtak
-import no.nav.familie.klage.brev.ef.domain.Brevmottaker
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerOrganisasjon
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerPerson
-import no.nav.familie.klage.brev.ef.domain.Brevmottakere
-import no.nav.familie.klage.brev.ef.domain.MottakerRolle
+import no.nav.familie.klage.brev.ef.Brevmottaker
+import no.nav.familie.klage.brev.ef.BrevmottakerOrganisasjon
+import no.nav.familie.klage.brev.ef.BrevmottakerPerson
+import no.nav.familie.klage.brev.ef.Brevmottakere
+import no.nav.familie.klage.brev.ef.MottakerRolle
 import no.nav.familie.klage.fagsak.domain.Fagsak
 import no.nav.familie.klage.fagsak.domain.tilYtelse
 import no.nav.familie.klage.infrastruktur.config.LenkeConfig
