+++ src/test/kotlin/no/nav/familie/klage/brev/baks/BaksBrevOppretterTest.kt
@@ -6,7 +6,7 @@ import io.mockk.mockkObject
 import io.mockk.unmockkObject
 import no.nav.familie.klage.behandling.BehandlingService
 import no.nav.familie.klage.behandling.domain.StegType
-import no.nav.familie.klage.brev.FamilieDokumentClient
+import no.nav.familie.klage.brev.felles.FamilieDokumentClient
 import no.nav.familie.klage.infrastruktur.exception.Feil
 import no.nav.familie.klage.infrastruktur.sikkerhet.SikkerhetContext
 import no.nav.familie.klage.testutil.DomainUtil
