+++ src/test/kotlin/no/nav/familie/klage/brev/felles/AvsnittRepositoryTest.kt
@@ -1,7 +1,5 @@
-package no.nav.familie.klage.brev
+package no.nav.familie.klage.brev.felles
 
-import no.nav.familie.klage.brev.ef.AvsnittRepository
-import no.nav.familie.klage.brev.ef.domain.Avsnitt
 import no.nav.familie.klage.infrastruktur.config.OppslagSpringRunnerTest
 import no.nav.familie.klage.repository.findByIdOrThrow
 import no.nav.familie.klage.testutil.DomainUtil.behandling
