+++ src/test/kotlin/no/nav/familie/klage/distribusjon/JournalførBrevTaskTest.kt
@@ -6,14 +6,14 @@ import io.mockk.mockk
 import io.mockk.verify
 import io.mockk.verifyOrder
 import no.nav.familie.klage.behandling.BehandlingService
+import no.nav.familie.klage.brev.ef.Brev
 import no.nav.familie.klage.brev.ef.BrevService
-import no.nav.familie.klage.brev.ef.domain.Brev
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerOrganisasjon
-import no.nav.familie.klage.brev.ef.domain.BrevmottakerPerson
-import no.nav.familie.klage.brev.ef.domain.Brevmottakere
-import no.nav.familie.klage.brev.ef.domain.BrevmottakereJournalpost
-import no.nav.familie.klage.brev.ef.domain.BrevmottakereJournalposter
-import no.nav.familie.klage.brev.ef.domain.MottakerRolle
+import no.nav.familie.klage.brev.ef.BrevmottakerOrganisasjon
+import no.nav.familie.klage.brev.ef.BrevmottakerPerson
+import no.nav.familie.klage.brev.ef.Brevmottakere
+import no.nav.familie.klage.brev.ef.BrevmottakereJournalpost
+import no.nav.familie.klage.brev.ef.BrevmottakereJournalposter
+import no.nav.familie.klage.brev.ef.MottakerRolle
 import no.nav.familie.klage.distribusjon.ef.DistribuerBrevTask
 import no.nav.familie.klage.distribusjon.ef.JournalførBrevTask
 import no.nav.familie.klage.felles.domain.Fil
